requirements.txt


openai
tiktoken
OpenAIAuth==0.3.2
httpx
prompt-toolkit

# haigf
# PySide2  # for haigf
pillow  # for haigf
pyyaml  # for haigf
# pyqtgraph  # for haigf

damei
numpy # for damei
easydict  # for damei

mysqlclient # for sql
pymysql # for sql
Flask-SQLAlchemy  # for sql


# python-ldap  # for sso-auth, 需要`conda install python-ldap`处理依赖关系



